+++ molecularnodes/style/__init__.py
@@ -1,0 +1,1 @@
+from .interface import add_style_branch
